requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
asyncpg==0.29.0
blinker==1.8.2
click==8.1.7
Flask==3.0.3
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
typing_extensions==4.12.2
Werkzeug==3.0.3
zipp==3.19.2
